Modified Card.js, EventPage.js & MyEvents.js to include event listener
