+++ src/components/EventPage.js
@@ -1,11 +1,37 @@
 import React from 'react';
 import { useLocation } from 'react-router-dom';
 import Nav from './Nav';
+import { getDatabase, ref, push } from 'firebase/database';
+import MyEvents from './MyEvents';
 
 
 export default function EventPage(props) {
-  const location = useLocation();
-  const {title, eventLocation, description, image, date} = location.state;
+  const loc = useLocation();
+  const {title, location, description, startTime, image, alt, date} = loc.state;
+
+  // handle registration
+  function handleRegister(e) {
+    e.preventDefault();
+    console.log('You registered for this event.');
+
+    const db = getDatabase();
+    const dbRef = ref(db, 'my-events');
+
+    if (e) {
+        const regEvent = {
+            title: {title},
+            location: {location},
+            description: {description},
+            startTime: {startTime},
+            image: {image},
+            alt: {alt}
+        }
+
+        push(regEvent, dbRef)
+        .catch((error) => console.log('Error: ', error));
+    }
+    <MyEvents />
+};
 
   return(
     <div>
@@ -17,10 +43,10 @@ export default function EventPage(props) {
       <div className="event_info">
         <div className="Titles">
           <h5><div className="event_details"><strong className="colorTitle">Time: </strong> {date}</div></h5>
-          <h5><div className="event_details"><strong className="colorTitle">Location: </strong>{eventLocation}</div></h5>
+          <h5><div className="event_details"><strong className="colorTitle">Location: </strong>{location}</div></h5>
         </div>
         <h4><div className="event_description">{description}</div></h4>
-        <a href="#" className="btn btn-dark register">Register</a>
+        <button type="submit" onClick={handleRegister} className="btn btn-dark register">Register</button>
         </div>
       </main>
     </div>
